Error handling improved to deleting/adding reactions

+++ app/socket_helpers/handle_delete_reaction_helper.py
@@ -3,7 +3,10 @@ from app.models import db, Reaction
 
 def handle_delete_reaction_helper(data):
     reaction_id = data.get("id")
-    reaction = db.session.query(Reaction).get(reaction_id)
+    reaction = Reaction.query.get(reaction_id)
+
+    if not reaction:
+        return { "error": "Reaction not found" }
 
     db.session.delete(reaction)
     db.session.commit()
